user completed

+++ db/migrate/20230705014054_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration[7.0]
     create_table :users do |t|
       t.string :username
       t.integer :age
-      t.integer :weight
+      t.float :weight
       t.integer :height
       t.boolean :male
